Allows passing nodes to the time prop of Timeline
If we want a link on the time for example, this can be useful.
